+++ src/Linear/V1.hs
@@ -88,6 +88,7 @@ import Linear.Metric
 import Linear.Epsilon
 import Linear.Vector
 import Prelude hiding (sum)
+import System.Random
 
 import qualified Data.Vector.Generic.Mutable as M
 import qualified Data.Vector.Generic as G
@@ -391,6 +392,13 @@ instance Binary a => Binary (V1 a) where
 instance Serialize a => Serialize (V1 a) where
   put = serializeWith Cereal.put
   get = deserializeWith Cereal.get
+
+instance Random a => Random (V1 a) where
+  random g = case random g of (a, g') -> (V1 a, g')
+  randoms g = V1 <$> randoms g
+  randomR (V1 a, V1 b) g = case randomR (a, b) g of (a', g') -> (V1 a', g')
+  randomRs (V1 a, V1 b) g = V1 <$> randomRs (a, b) g
+  randomIO = V1 <$> randomIO
 
 #if (MIN_VERSION_transformers(0,5,0)) || !(MIN_VERSION_transformers(0,4,0))
 instance Eq1 V1 where